src/lib.rs
use rayon::prelude::*;
use serde::Serialize;
use std::error::Error;
use std::ffi::OsStr;
use std::fs;
use std::path::Path;

mod ffi;

#[cfg(test)]
mod tests;

#[derive(Serialize)]
pub struct DiskItem {
    pub name: String,
    pub disk_size: u64,
    pub children: Option<Vec<DiskItem>>,
}

impl DiskItem {
    /// Analyzes provided path and returns tree structure of analyzed DiskItems
    pub fn from_analyze(
        path: &Path,
        apparent: bool,
        root_dev: u64,
    ) -> Result<Self, Box<dyn Error>> {
        let name = path
            .file_name()
            .unwrap_or(&OsStr::new("."))
            .to_string_lossy()
            .to_string();

        let file_info = FileInfo::from_path(path, apparent)?;

        match file_info {
            FileInfo::Directory { volume_id } => {
                if volume_id != root_dev {
                    return Err("Filesystem boundary crossed".into());
                }

                let sub_entries = fs::read_dir(path)?
                    .filter_map(Result::ok)
                    .collect::<Vec<_>>();

                let mut sub_items = sub_entries
                    .par_iter()
                    .filter_map(|entry| {
                        DiskItem::from_analyze(&entry.path(), apparent, root_dev).ok()
                    })
                    .collect::<Vec<_>>();

                sub_items.sort_unstable_by(|a, b| a.disk_size.cmp(&b.disk_size).reverse());

                Ok(DiskItem {
                    name,
                    disk_size: sub_items.iter().map(|di| di.disk_size).sum(),
                    children: Some(sub_items),
                })
            }
            FileInfo::File { size, .. } => Ok(DiskItem {
                name,
                disk_size: size,
                children: None,
            }),
        }
    }
}

pub enum FileInfo {
    File { size: u64, volume_id: u64 },
    Directory { volume_id: u64 },
}

impl FileInfo {
    #[cfg(unix)]
    pub fn from_path(path: &Path, apparent: bool) -> Result<Self, Box<dyn Error>> {
        use std::os::unix::fs::MetadataExt;

        let md = path.symlink_metadata()?;
        if md.is_dir() {
            Ok(FileInfo::Directory {
                volume_id: md.dev(),
            })
        } else {
            let size = if apparent {
                md.blocks() * 512
            } else {
                md.len()
            };
            Ok(FileInfo::File {
                size,
                volume_id: md.dev(),
            })
        }
    }

    #[cfg(windows)]
    pub fn from_path(path: &Path, apparent: bool) -> Result<Self, Box<dyn Error>> {
        use crate::ffi::get_file_information_by_handle_ex;
        use std::convert::TryInto;
        use winapi::um::fileapi::FILE_ID_INFO;
        use winapi::um::fileapi::FILE_STANDARD_INFO;
        use winapi::um::winnt::LARGE_INTEGER;
        use winapi_util::Handle;

        let h = Handle::from_path_any(path)?;
        let std_info: FILE_STANDARD_INFO = get_file_information_by_handle_ex(&h)?;
        // That's unfortunate that we have to make second syscall just to know volume serial number
        let id_info: FILE_ID_INFO = get_file_information_by_handle_ex(&h)?;
        // If we decide to skip symlinks, we would also need FILE_ATTRIBUTE_TAG_INFO struct.

        if std_info.Directory != 0 {
            Ok(FileInfo::Directory {
                volume_id: id_info.VolumeSerialNumber,
            })
        } else {
            let size: LARGE_INTEGER = if apparent {
                std_info.AllocationSize
            } else {
                std_info.EndOfFile
            };

            let size = ffi::read_large_integer(size);

            Ok(FileInfo::File {
                size: size.try_into()?, // it is i64 on windows
                volume_id: id_info.VolumeSerialNumber,
            })
        }
    }
}
